requirements.txt
# Core Flask Framework
Flask==2.3.3
Werkzeug==2.3.7

# Database
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.39
PyMySQL==1.1.0
Flask-Migrate==4.0.5

# Authentication & Security
Flask-JWT-Extended==4.5.3
bcrypt==4.0.1
Flask-WTF==1.1.1

# Serialization & Validation
flask-marshmallow==1.2.0
marshmallow==3.21.3
marshmallow-sqlalchemy==0.29.0
email-validator==2.0.0
WTForms==3.0.1

# Rate Limiting & Performance
Flask-Limiter==3.5.0

# Environment & Configuration
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Chatbot AI Dependencies
google-genai>=1.0.0               # For Google Gemini API integration
requests==2.31.0                  # For HTTP requests

# Optional Dependencies for Enhanced Features
# Uncomment if needed:
# redis==4.6.0                    # For Redis-based rate limiting
# celery==5.3.1                   # For background tasks
# flask-caching==2.1.0            # For caching
# pillow==10.0.0                  # For image processing
